requirements.txt
discord.py
asyncio
ntscraper
datetime
tqdm